[Starts #161642802]Add tests for user endpoint
